requirements.txt
spotipy>=2.23.0
pypresence>=4.3.0
websockets>=15.0.1
python-dotenv>=1.0.0